# docs/api/auth-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth Center API",
    "version": "1.5.0"
  },
  "servers": [
    {
      "url": "https://taxindo.ai"
    }
  ],
  "paths": {
...
    "/api/billing/users": {
      "get": {
        "summary": "List users (Billing/Admin)",
        "security": [
          {
            "billingApiKey": []
          },
          {
            "billingBearer": []
          },
          {
            "integrationApiKey": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FREE",
                "BASIC",
                "PLUS",
                "MAX"
              ]
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "USER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tpc_session"
      },
      "integrationApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "billingApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "billingBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "BillingUsersResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "name": {
                  "type": "string",
                  "nullable": true
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "ADMIN",
                    "USER"
                  ]
                },
                "plan": {
                  "type": "string",
                  "enum": [
                    "FREE",
                    "BASIC",
                    "PLUS",
                    "MAX"
                  ]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
    }
  }
}